+++ psycop/common/model_training_v2/trainer/task/base_task.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Protocol
+from typing import TYPE_CHECKING, Protocol, runtime_checkable
 
 if TYPE_CHECKING:
     import pandas as pd
@@ -11,6 +11,7 @@ if TYPE_CHECKING:
     )
 
 
+@runtime_checkable
 class BaselineTask(Protocol):
     def train(self, x: PolarsFrame, y: PolarsFrame) -> None:
         """Train the model"""