+++ psycop/common/model_training_v2/trainer/task/binary_classification/binary_classification_pipeline.py
@@ -1,27 +1,24 @@
-from collections.abc import Sequence
-
 import pandas as pd
 import polars as pl
 from sklearn.pipeline import Pipeline
 
+from psycop.common.model_training_v2.config.baseline_registry import BaselineRegistry
 from psycop.common.model_training_v2.trainer.preprocessing.polars_frame import (
     PolarsFrame,
-)
-from psycop.common.model_training_v2.trainer.task.model_step import (
-    ModelStep,
 )
 
 PredProbaSeries = pd.Series  # name should be "y_hat_probs", series of floats
 
 
+@BaselineRegistry.task_pipelines.register("binary_classification_pipeline")
 class BinaryClassificationPipeline:
-    def __init__(self, steps: Sequence[ModelStep]):
-        self.pipe = Pipeline(steps=steps)
+    def __init__(self, sklearn_pipe: Pipeline):
+        self.pipe = sklearn_pipe
 
     def fit(self, x: PolarsFrame, y: pl.Series) -> None:
         if isinstance(x, pl.LazyFrame):
             x = x.collect()
-        self.pipe.fit(x.to_pandas(), y)
+        self.pipe.fit(X=x.to_pandas(), y=y)
 
     def predict_proba(self, x: PolarsFrame) -> PredProbaSeries:
         """Returns the predicted probabilities of the `1`