+++ psycop/common/model_training_v2/trainer/task/binary_classification/binary_metrics/binary_auroc.py
@@ -4,6 +4,7 @@ from typing import TYPE_CHECKING
 
 from sklearn.metrics import roc_auc_score
 
+from psycop.common.model_training_v2.config.baseline_registry import BaselineRegistry
 from psycop.common.model_training_v2.trainer.task.base_metric import (
     CalculatedMetric,
 )
@@ -19,7 +20,11 @@ if TYPE_CHECKING:
     )
 
 
+@BaselineRegistry.metrics.register("binary_auroc")
 class BinaryAUROC(BinaryMetric):
+    def __init__(self) -> None:
+        pass
+
     def calculate(
         self,
         y_true: pd.Series[int],