+++ psycop/common/model_training_v2/trainer/task/binary_classification/test_binary_classification_pipeline.py
@@ -1,8 +1,7 @@
-from collections.abc import Sequence
-
 import pandas as pd
 import polars as pl
 import pytest
+from sklearn.pipeline import Pipeline
 
 from psycop.common.model_training_v2.trainer.preprocessing.polars_frame import (
     PolarsFrame,
@@ -13,27 +12,24 @@ from psycop.common.model_training_v2.trainer.task.binary_classification.binary_c
 from psycop.common.model_training_v2.trainer.task.estimator_steps.logistic_regression import (
     logistic_regression_step,
 )
-from psycop.common.model_training_v2.trainer.task.model_step import (
-    ModelStep,
-)
 
 
 @pytest.mark.parametrize(
-    ("steps", "x", "y"),
+    ("pipe", "x", "y"),
     [
         (
-            [logistic_regression_step()],
+            Pipeline([logistic_regression_step()]),
             pl.DataFrame({"x": [1, 2, 3, 4]}),
             pl.Series([0, 0, 1, 1]),
         ),
     ],
 )
 def test_binary_classification_pipeline(
-    steps: Sequence[ModelStep],
+    pipe: Pipeline,
     x: PolarsFrame,
     y: pl.Series,
 ):
-    pipeline = BinaryClassificationPipeline(steps=steps)
+    pipeline = BinaryClassificationPipeline(sklearn_pipe=pipe)
     pipeline.fit(x=x, y=y)
 
     y_hat_probs = pipeline.predict_proba(x=x)