+++ psycop/common/model_training_v2/trainer/task/estimator_steps/logistic_regression.py
@@ -2,11 +2,13 @@ from typing import Literal
 
 from sklearn.linear_model import LogisticRegression
 
+from psycop.common.model_training_v2.config.baseline_registry import BaselineRegistry
 from psycop.common.model_training_v2.trainer.task.model_step import (
     ModelStep,
 )
 
 
+@BaselineRegistry.estimator_steps.register("logistic_regression")
 def logistic_regression_step(
     penalty: Literal["l1", "l2", "elasticnet"] = "elasticnet",
     solver: Literal[