+++ psycop/common/model_training_v2/trainer/task/test_binary_classification.py
@@ -1,6 +1,7 @@
 import polars as pl
 import pytest
 from polars.testing import assert_series_equal
+from sklearn.pipeline import Pipeline
 
 from psycop.common.model_training_v2.trainer.preprocessing.polars_frame import (
     PolarsFrame,
@@ -26,7 +27,9 @@ from psycop.common.model_training_v2.trainer.task.estimator_steps.logistic_regre
     ("pipe", "main_metric", "x", "y", "main_metric_expected"),
     [
         (
-            BinaryClassificationPipeline(steps=[logistic_regression_step()]),
+            BinaryClassificationPipeline(
+                sklearn_pipe=Pipeline([logistic_regression_step()]),
+            ),
             BinaryAUROC(),
             pl.DataFrame({"x": [1, 1, 2, 2], "uuid": [1, 2, 3, 4]}),
             pl.DataFrame({"y": [0, 0, 1, 1]}),
@@ -42,7 +45,7 @@ def test_binary_classification(
     main_metric_expected: float,
 ):
     binary_classification_problem = BinaryClassification(
-        pipe=pipe,
+        task_pipe=pipe,
         main_metric=main_metric,
         pred_time_uuid_col_name="uuid",
     )